Fixes have been made
